delete when status = draft

+++ Backend/cart-service/src/main/java/com/example/cart_service/service/implement/OrderServiceImpl.java
@@ -48,6 +48,11 @@ public class OrderServiceImpl implements OrderService {
     @Override
     @Transactional
     public void delete(Integer id) {
+        Order order = orderRepository.findById(id)
+                .orElseThrow(() -> new RuntimeException("Order not found"));
+        if(!"DRAFT".equals(order.getStatus())) {
+            throw new RuntimeException("Only DRAFT orders can be deleted");
+        }
         orderRepository.deleteById(id);
     }
